update Corona API get increased case

+++ src/main/java/net/qsef1256/capstone2022server/api/corona/CoronaAPI.java
@@ -1,6 +1,7 @@
 package net.qsef1256.capstone2022server.api.corona;
 
 import lombok.extern.slf4j.Slf4j;
+
 import net.qsef1256.capstone2022server.api.APIConnector;
 import net.qsef1256.capstone2022server.api.enums.APICode;
 import net.qsef1256.capstone2022server.api.enums.ContentType;
@@ -27,11 +28,13 @@ public class CoronaAPI {
 
     static {
         new CoronaAPI().update();
-        DiaScheduler.executePerDay(() -> new CoronaAPI().update());
+        DiaScheduler.executePerTime(() -> new CoronaAPI().update(), 12, 0, 0); // 정각 12시 업데이트
     }
 
     /**
      * 확진자와 사망자 기록을 업데이트 합니다.
+     *
+     * <p>오늘자 데이터가 없을 경우 업데이트에 실패할 수 있습니다. 이 경우 어제의 데이터를 놔두게 됩니다.</p>
      */
     public void update() {
         CoronaEntity data = getData();
@@ -42,10 +45,11 @@ public class CoronaAPI {
 
         System.out.println("Corona data updated called");
         String key = PropertiesUtil.loadFile("key").getProperty("corona.token");
-        String date = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyyMMdd"));
+        String startDate = LocalDateTime.now().minusDays(1).format(DateTimeFormatter.ofPattern("yyyyMMdd"));
+        String endDate = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyyMMdd"));
 
         String url = "http://openapi.data.go.kr/openapi/service/rest/Covid19/getCovid19InfStateJson"
-                + "?serviceKey=" + key + "&startCreateDt=" + date;
+                + "?serviceKey=" + key + "&startCreateDt=" + startDate + "&endCreateDt=" + endDate;
 
         try (APIConnector connector = new APIConnector()) {
             Document document = XmlUtil.parse(connector.getResult(url, ContentType.XML));
@@ -55,17 +59,39 @@ public class CoronaAPI {
             if (apiCode != APICode.NORMAL_SERVICE)
                 throw new IllegalArgumentException("공공 API 와 연동하는데 실패했습니다: " + (apiCode != null ? apiCode.getDisplay() : null));
 
-            Element item = document.getRootElement()
-                    .element("body").element("items").elements().get(0);
-            long deathCnt = Long.parseLong(item.element("deathCnt").getText());
-            long decideCnt = Long.parseLong(item.element("decideCnt").getText());
-            LocalDateTime updateTime = LocalDateTime.now();
+            List<Element> items = document.getRootElement().element("body").element("items").elements();
 
-            dao.open();
-            dao.deleteAll();
-            dao.saveAllAndClose(Collections.singleton(new CoronaEntity(deathCnt, decideCnt, updateTime)));
+            if (items.size() == 2) {
+                Element startItem, endItem;
+
+                if (items.get(0).element("stateDt").getText().equals(startDate)) {
+                    startItem = items.get(0);
+                    endItem = items.get(1);
+                } else {
+                    startItem = items.get(1);
+                    endItem = items.get(0);
+                }
+
+                long startDeathCnt = Long.parseLong(startItem.element("deathCnt").getText());
+                long startDecideCnt = Long.parseLong(startItem.element("decideCnt").getText());
+
+                long endDeathCnt = Long.parseLong(endItem.element("deathCnt").getText());
+                long endDecideCnt = Long.parseLong(endItem.element("decideCnt").getText());
+
+                long addDeath = endDeathCnt - startDeathCnt;
+                long addDecide = endDecideCnt - startDecideCnt;
+
+                LocalDateTime updateTime = LocalDateTime.now();
+
+                dao.open();
+                dao.deleteAll();
+                dao.saveAllAndClose(Collections.singleton(new CoronaEntity(endDeathCnt, endDecideCnt, addDeath, addDecide, updateTime)));
+            } else {
+                log.warn("Failed to get Today's COVID-19 data");
+            }
         } catch (IOException | DocumentException | RuntimeException e) {
             e.printStackTrace();
+            throw new RuntimeException("Failed to update COVID-19 data");
         }
 
     }
